pyproject.toml
[project]
name = "djangoblog"
version = "0.1.0"
description = ""
authors = [
    {name = "sonjack",email = "823266085@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-allauth (>=65.4.1,<66.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "django-anymail[mailgun] (>=12.0,<13.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise[brotli] (>=6.9.0,<7.0.0)",

]

[tool.poetry.dependencies]
django = "5.0.1"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
django-browser-reload = "^1.18.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
playwright = "^1.50.0"
pytest-playwright = "^0.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
